Use Blake2bHasher
